Update Box3D to 9e5a4cd
Box3D moves from 3fc20f5 to 9e5a4cd, the six upstream commits between
them. The bindings are regenerated, abi/native-layout.json is re-recorded
from the C compiler, and every changed struct mirror follows its header:
b3MeshDef gains a size_t stride and clockWiseWinding, b3PlaneResult gains
triangle, child and material indices, b3BodyTOIResult is new, the dynamic
tree is rewritten around 32-byte paired nodes and a separate proxy array,
and hulls, meshes and height fields carry 64-bit hashes.

The idiomatic layer only gains members, and package validation against
the last release reports no difference for it:

- MeshOptions.ClockwiseWinding, and CollisionMesh.FromTriangles<TVertex>
  for positions read out of an interleaved vertex buffer.
- CharacterContact.TriangleIndex, ChildIndex and MaterialIndex.
- CharacterMover.TimeOfImpact, for a moving body striking a character.
- Joint.IsAwake, and SafetyFactor on BodyDefinition and Body.

Box3D.NET.Native breaks where the C declarations broke: b3GetByteCount
returns int64, the allocator callbacks take size_t, and the tree, hash,
counter and profile fields moved. These are recorded in
CompatibilitySuppressions.xml. b3RecPlayer_Create and _Destroy were pure
renames upstream, so the old names stay as obsolete forwarders.

Shape casts now pass through the back of mesh and height field triangles,
as ray casts always have. The tests that pin this fail against the
previous Box3D. A NaN smuggled in through the native layer no longer
reaches bodies it is not touching, so the fuzz test that justifies the
validation now measures contamination through contact instead.

+++ src/Box3D.NET.Tests/GeometryOwnershipTests.cs
@@ -44,9 +44,9 @@ public class GeometryOwnershipTests
         // measuring from cold would report that as a leak.
         operation();
 
-        int before = B3.b3GetByteCount();
+        long before = B3.b3GetByteCount();
         operation();
-        int after = B3.b3GetByteCount();
+        long after = B3.b3GetByteCount();
 
         Assert.True(
             before == after,
